requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt
importlib-metadata==4.6.1
Markdown==3.3.4
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.5.0
djangorestframework-jwt
channels
redis
channels-redis
psycopg2-binary==2.9.1
django-xmpp
